chore(eslint): enhanced restricted imports
