+++ pkg/api/api/browser/projects/_projectId/desks/index.ts
@@ -1,4 +1,5 @@
-import type { ApiDesk, ProjectId } from '@violet/api/types'
+import type { ApiDesk } from '@violet/lib/types/api'
+import type { ProjectId } from '@violet/lib/types/branded'
 
 export type Methods = {
   get: {
