+++ pkg/api/api/browser/works/_workId@string/revisions/_revisionId@string/messages/_messageId@string/replies/index.ts
@@ -1,4 +1,4 @@
-import type { ApiReply } from '@violet/api/types'
+import type { ApiReply } from '@violet/lib/types/api'
 
 export type Methods = {
   post: {
