+++ pkg/api/api/browser/works/_workId@string/revisions/_revisionId@string/messages/controller.ts
@@ -1,5 +1,5 @@
 import { getMessages } from '@violet/api/src/service/streamBar'
-import type { RevisionId } from '@violet/api/types'
+import type { RevisionId } from '@violet/lib/types/branded'
 import { defineController } from './$relay'
 
 export default defineController(() => ({
