+++ pkg/api/api/browser/works/_workId@string/revisions/_revisionId@string/messages/index.ts
@@ -1,4 +1,5 @@
-import type { ApiMessage, RevisionId } from '@violet/api/types'
+import type { ApiMessage } from '@violet/lib/types/api'
+import type { RevisionId } from '@violet/lib/types/branded'
 
 export type Methods = {
   get: {
