+++ pkg/api/src/service/s3.ts
@@ -6,8 +6,8 @@ import {
   S3Client,
 } from '@aws-sdk/client-s3'
 import { getSignedUrl } from '@aws-sdk/s3-request-presigner'
-import type { S3SaveWorksPath } from '@violet/api/types'
 import { fileTypes } from '@violet/def/constants'
+import type { S3SaveWorksPath } from '@violet/lib/types/branded'
 import type { MultipartFile } from 'fastify-multipart'
 import { depend } from 'velona'
 import envValues from '../utils/envValues'
