+++ pkg/api/src/types/index.d.ts
@@ -1,0 +1,1 @@
+export type Sample = { a: string }
