+++ pkg/api/src/utils/s3.ts
@@ -1,4 +1,4 @@
-import type { DeskId, ProjectId, RevisionId, S3SaveWorksPath } from '@violet/api/types'
+import type { DeskId, ProjectId, RevisionId, S3SaveWorksPath } from '@violet/lib/types/branded'
 
 export const createS3SaveWorksPath = (props: {
   projectId: ProjectId
