--- pkg/api/types/index.d.ts
@@ -1,13 +1,0 @@
-export * from './api'
-export * from './branded'
-export * from './client'
-
-export type UserInfo = {
-  id: string
-  name: string
-  icon: string
-}
-
-export type AuthHeader = {
-  authorization: string
-}
