--- pkg/api/types/prisma-client.d.ts
@@ -1,1 +1,0 @@
-export * from '@prisma/client'
