+++ pkg/lib/types/branded.d.ts
