+++ pkg/web/src/pages/browser/[...pathes]/components/Explorer/WorkCell.tsx
@@ -1,4 +1,4 @@
-import type { BrowserWork } from '@violet/api/types'
+import type { BrowserWork } from '@violet/web/src/types/browser'
 import { getWorkFullName } from '@violet/web/src/utils'
 import { CellName } from './CellName'
