+++ pkg/web/src/pages/browser/[...pathes]/components/ProjectBar.tsx
@@ -1,11 +1,12 @@
-import type { BrowserProject, ProjectId } from '@violet/api/types'
-import { alphaLevel, colors, fontSizes } from '@violet/web/src//utils/constants'
+import type { ProjectId } from '@violet/lib/types/branded'
 import { PlusIcon } from '@violet/web/src/components/atoms/PlusIcon'
 import { Spacer } from '@violet/web/src/components/atoms/Spacer'
 import { Modal } from '@violet/web/src/components/molecules/Modal'
 import { pagesPath } from '@violet/web/src/utils/$path'
+import { alphaLevel, colors, fontSizes } from '@violet/web/src/utils/constants'
 import Link from 'next/link'
 import { useState } from 'react'
+import type { BrowserProject } from 'src/types/browser'
 import styled from 'styled-components'
 import { ProjectNameInput } from './ProjectNameInput'
