+++ pkg/web/src/pages/browser/[...pathes]/components/StreamBar/MessageCell.tsx
@@ -1,7 +1,8 @@
-import type { BrowserMessage, MessageId } from '@violet/api/types'
-import { Spacer } from '@violet/web/src//components/atoms/Spacer'
-import { alphaLevel, colors } from '@violet/web/src//utils/constants'
+import type { MessageId } from '@violet/lib/types/branded'
+import { Spacer } from '@violet/web/src/components/atoms/Spacer'
+import { alphaLevel, colors } from '@violet/web/src/utils/constants'
 import React from 'react'
+import type { BrowserMessage } from 'src/types/browser'
 import styled from 'styled-components'
 import { MessageHeader } from './MessageHeader'
 import { ReplyInputForm } from './ReplyInputForm'
