+++ pkg/web/src/pages/browser/[...pathes]/components/StreamBar/ReplyInputForm.tsx
@@ -1,5 +1,5 @@
-import { Spacer } from '@violet/web/src//components/atoms/Spacer'
-import { alphaLevel, colors } from '@violet/web/src//utils/constants'
+import { Spacer } from '@violet/web/src/components/atoms/Spacer'
+import { alphaLevel, colors } from '@violet/web/src/utils/constants'
 import React, { useState } from 'react'
 import { PencilIcon } from 'src/components/atoms/PencilIcon'
 import styled from 'styled-components'
