+++ pkg/web/src/pages/browser/[...pathes]/components/TabBar.tsx
@@ -1,9 +1,9 @@
-import type { BrowserProject, ProjectApiData } from '@violet/api/types'
 import { Spacer } from '@violet/web/src/components/atoms/Spacer'
 import { getWorkFullName } from '@violet/web/src/utils'
 import { pagesPath } from '@violet/web/src/utils/$path'
 import { alphaLevel, colors } from '@violet/web/src/utils/constants'
 import Link from 'next/link'
+import type { BrowserProject, ProjectApiData } from 'src/types/browser'
 import styled from 'styled-components'
 import { ExtIcon } from './ExtIcon'
 import { SelectableStyle } from './SelectableStyle'
