+++ pkg/web/src/pages/browser/[...pathes]/index.page.tsx
@@ -1,6 +1,6 @@
-import type { BrowserRevision } from '@violet/api/types'
 import { Fetching } from '@violet/web/src/components/organisms/Fetching'
 import { useMemo } from 'react'
+import type { BrowserRevision } from 'src/types/browser'
 import styled from 'styled-components'
 import { EmptyWork } from './components/EmptyWork'
 import { Explorer } from './components/Explorer'
